split API into tiers
