+++ pyclesperanto/_tier1.py
@@ -1,11 +1,11 @@
 from ._image import Image
-from ._device import cleDevice, get_device
+from ._device import Device
 from ._decorators import plugin_function
 
 
 @plugin_function
 def absolute(
-    input_image: Image, output_image: Image = None, device: cleDevice = None
+    input_image: Image, output_image: Image = None, device: Device = None
 ) -> Image:
     """absolute
 
@@ -22,9 +22,6 @@ def absolute(
     """
     from ._pyclesperanto import _AbsoluteKernel_Call as op
 
-    if device is None:
-        device = get_device()
-
     op(device, input_image, output_image)
     return output_image
 
@@ -36,13 +33,10 @@ def gaussian_blur(
     sigma_x: float = 0,
     sigma_y: float = 0,
     sigma_z: float = 0,
-    device: cleDevice = None,
+    device: Device = None,
 ) -> Image:
 
     from ._pyclesperanto import _GaussianBlurKernel_Call as op
-
-    if device is None:
-        device = get_device()
 
     op(
         device,
@@ -59,13 +53,10 @@ def gaussian_blur(
 def maximum_z_projection(
     input_image: Image,
     output_image: Image = None,
-    device: cleDevice = None,
+    device: Device = None,
 ) -> Image:
 
     from ._pyclesperanto import _MaximumZProjectionKernel_Call as op
-
-    if device is None:
-        device = get_device()
 
     op(device, input_image, output_image)
     return output_image
@@ -75,13 +66,10 @@ def maximum_z_projection(
 def maximum_y_projection(
     input_image: Image,
     output_image: Image = None,
-    device: cleDevice = None,
+    device: Device = None,
 ) -> Image:
 
     from ._pyclesperanto import _MaximumYProjectionKernel_Call as op
-
-    if device is None:
-        device = get_device()
 
     op(device, input_image, output_image)
     return output_image
@@ -91,46 +79,10 @@ def maximum_y_projection(
 def maximum_x_projection(
     input_image: Image,
     output_image: Image = None,
-    device: cleDevice = None,
+    device: Device = None,
 ) -> Image:
 
     from ._pyclesperanto import _MaximumXProjectionKernel_Call as op
-
-    if device is None:
-        device = get_device()
-
-    op(device, input_image, output_image)
-    return output_image
-
-
-
-@plugin_function
-def threshold_otsu(
-    input_image: Image,
-    output_image: Image = None,
-    device: cleDevice = None,
-) -> Image:
-
-    from ._pyclesperanto import _ThresholdOtsuKernel_Call as op
-
-    if device is None:
-        device = get_device()
-
-    op(device, input_image, output_image)
-    return output_image
-
-
-@plugin_function
-def connect_component_labeling_box(
-    input_image: Image,
-    output_image: Image = None,
-    device: cleDevice = None,
-) -> Image:
-
-    from ._pyclesperanto import _ConnectedComponentLabelingBoxKernel_Call as op
-
-    if device is None:
-        device = get_device()
 
     op(device, input_image, output_image)
     return output_image
